Implemented Product commands

+++ src/Shop.Domain/ProductAggregate/Value Objects/Score.cs
@@ -5,18 +5,18 @@ namespace Shop.Domain.ProductAggregate.Value_Objects;
 
 public class Score : BaseValueObject
 {
-    public float Value { get; private set; }
+    public int Value { get; private set; }
 
     private const int MinimumScore = 0;
     private const int MaximumScore = 5;
     
-    public Score(float scoreValue)
+    public Score(int scoreValue)
     {
         Guard(scoreValue);
         Value = scoreValue;
     }
 
-    private void Guard(float scoreValue)
+    private void Guard(int scoreValue)
     {
         OutOfRangeValueDomainException.CheckRange(MinimumScore, MaximumScore, scoreValue, nameof(scoreValue));
     }
